adds SWEET dataset prefix
"A bit more context on the types of ‘things’ you are trying to model would be very helpful. Are we talking about Scientific Datasets? If so, it would be my opinion, that http://schema.org/Dataset is certainly more descriptive than http://sweetontology.net/reprDataProduct/Dataset and can therefore be considered as a subclass of." -- Lewis McGibbney